backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
psycopg2-binary
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pillow
tensorflow
python-dotenv
python-decouple
